% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cities}
\alias{cities}
\title{Bay Area Cities}
\format{
An sf object with 101 features and 5 fields:
\describe{
\item{city}{city name}
\item{county}{county name}
\item{source}{source}
\item{ugb}{unknown}
\item{acres}{area in acres}
\item{geometry}{feature geometry}
}
}
\source{
\url{https://earthworks.stanford.edu/catalog/stanford-vj593xs7263}
}
\usage{
cities
}
\description{
MULTIPOLYGON object of city limits in the San Francisco Bay Area
}
\keyword{datasets}
