% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties}
\alias{counties}
\title{Bay Area Counties}
\format{
An sf object with 9 features and 3 fields:
\describe{
\item{county}{county name}
\item{fipsstco}{fipsstco}
\item{objectid}{id}
\item{geometry}{feature geometry}
}
}
\source{
\url{https://earthworks.stanford.edu/catalog/ark28722-s7hs4j}
}
\usage{
counties
}
\description{
MULTIPOLYGON object of counties in the San Francisco Bay Area
}
\keyword{datasets}
