% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{highways}
\alias{highways}
\title{Bay Area Highways}
\format{
An sf object with 657 features and 14 fields:
\describe{
\item{fnode_}{fnode}
\item{tnode_}{tnode}
\item{lpoly_}{lpoly}
\item{rpoly_}{rpoly}
\item{length}{length in miles}
\item{st_hwy_dis}{st_hwy_dis}
\item{st_hwy_d_1}{st_hwy_d_1}
\item{rte}{route number - short}
\item{status}{status}
\item{funccl}{funccl}
\item{numlate}{number of lanes}
\item{accont}{accont}
\item{rsys}{rsys}
\item{label}{label}
\item{geometry}{feature geometry}
}
}
\source{
\url{https://earthworks.stanford.edu/catalog/stanford-vg822dt4563}
}
\usage{
highways
}
\description{
MULTIlINESTRING object of highways in the San Francisco Bay Area
}
\keyword{datasets}
