% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{streams}
\alias{streams}
\title{Bay Area Streams}
\format{
An sf object with 555 features and 3 fields:
\describe{
\item{length}{length in miles}
\item{pname}{p name}
\item{pnmae_lcas}{p name lcas}
\item{geometry}{feature geometry}
}
}
\source{
\url{https://earthworks.stanford.edu/catalog/ark28722-s7hs4j}
}
\usage{
streams
}
\description{
MULTILINESTRING object of streams in the San Francisco Bay Area
}
\keyword{datasets}
