% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water}
\alias{water}
\title{Bay Area Water}
\format{
An sf object with 1917 features and 0 fields:
\describe{
\item{geometry}{feature geometry}
}
}
\source{
\url{https://earthworks.stanford.edu/catalog/ark28722-s7fk57}
}
\usage{
water
}
\description{
POLYGON object of bodies of water in the San Francisco Bay Area
}
\keyword{datasets}
