% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{watersheds}
\alias{watersheds}
\title{Bay Area Watersheds}
\format{
An sf object with 21 features and 3 fields:
\describe{
\item{huname}{name}
\item{count}{count}
\item{acres}{acres}
\item{geometry}{feature geometry}
}
}
\source{
\url{https://earthworks.stanford.edu/catalog/ark28722-s7fk57}
}
\usage{
watersheds
}
\description{
MULTIPOLYGON object of watersheds in the San Francisco Bay Area
}
\keyword{datasets}
