% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zipcodes}
\alias{zipcodes}
\title{Bay Area Zipcodes}
\format{
An sf object with 187 features and 5 fields:
\describe{
\item{area}{area}
\item{length}{length}
\item{po_name}{po name}
\item{state}{state}
\item{zip}{zip code}
\item{geometry}{feature geometry}
}
}
\source{
\url{https://earthworks.stanford.edu/catalog/ark28722-s7888q}
}
\usage{
zipcodes
}
\description{
MULTIPOLYGON object of watersheds in the San Francisco Bay Area
}
\keyword{datasets}
